Fix merge forward duplicate tests
