+++ tests/pytests/unit/crypt/test_crypt_cryptodome.py
@@ -55,42 +55,3 @@ def test_gen_keys(tmp_path):
             salt.crypt.gen_keys(key_path, "keyname", 2048)
             assert open_priv_wb in m_open.calls
             assert open_pub_wb in m_open.calls
-
-
-@pytest.mark.slow_test
-def test_gen_keys_with_passphrase(tmp_path):
-
-    key_path = str(tmp_path / "keydir")
-    open_priv_wb = MockCall(os.path.join(key_path, "keyname.pem"), "wb+")
-    open_pub_wb = MockCall(os.path.join(key_path, "keyname.pub"), "wb+")
-
-    real_is_file = os.path.isfile
-
-    def is_file(path):
-        if path.startswith(str(tmp_path)):
-            return False
-        return real_is_file(path)
-
-    with patch.multiple(
-        os,
-        umask=MagicMock(),
-        chmod=MagicMock(),
-        access=MagicMock(return_value=True),
-    ):
-        salt.crypt.gen_keys(key_path, "keyname", 2048)
-        with patch("salt.utils.files.fopen", mock_open()) as m_open, patch(
-            "os.path.isfile", return_value=True
-        ):
-            result = salt.crypt.gen_keys(
-                key_path, "keyname", 2048, passphrase="password"
-            )
-            assert result == os.path.join(key_path, "keyname.pem")
-            assert open_priv_wb not in m_open.calls
-            assert open_pub_wb not in m_open.calls
-
-        with patch("salt.utils.files.fopen", mock_open()) as m_open, patch(
-            "salt.crypt.os.path.isfile", is_file
-        ):
-            salt.crypt.gen_keys(key_path, "keyname", 2048)
-            assert open_priv_wb in m_open.calls
-            assert open_pub_wb in m_open.calls
