+++ tests/pytests/unit/transport/test_zeromq.py
@@ -2,10 +2,8 @@ import ctypes
 import hashlib
 import logging
 import multiprocessing
-import os
 import threading
 import time
-import uuid
 
 import msgpack
 import pytest
@@ -260,98 +258,6 @@ def pki_dir(tmp_path):
     yield _pki_dir
 
 
-def test_master_uri():
-    """
-    test _get_master_uri method
-    """
-
-    m_ip = "127.0.0.1"
-    m_port = 4505
-    s_ip = "111.1.0.1"
-    s_port = 4058
-
-    m_ip6 = "1234:5678::9abc"
-    s_ip6 = "1234:5678::1:9abc"
-
-    with patch("salt.transport.zeromq.LIBZMQ_VERSION_INFO", (4, 1, 6)), patch(
-        "salt.transport.zeromq.ZMQ_VERSION_INFO", (16, 0, 1)
-    ):
-        # pass in both source_ip and source_port
-        assert (
-            salt.transport.zeromq._get_master_uri(
-                master_ip=m_ip, master_port=m_port, source_ip=s_ip, source_port=s_port
-            )
-            == f"tcp://{s_ip}:{s_port};{m_ip}:{m_port}"
-        )
-
-        assert (
-            salt.transport.zeromq._get_master_uri(
-                master_ip=m_ip6, master_port=m_port, source_ip=s_ip6, source_port=s_port
-            )
-            == f"tcp://[{s_ip6}]:{s_port};[{m_ip6}]:{m_port}"
-        )
-
-        # source ip and source_port empty
-        assert (
-            salt.transport.zeromq._get_master_uri(master_ip=m_ip, master_port=m_port)
-            == f"tcp://{m_ip}:{m_port}"
-        )
-
-        assert (
-            salt.transport.zeromq._get_master_uri(master_ip=m_ip6, master_port=m_port)
-            == f"tcp://[{m_ip6}]:{m_port}"
-        )
-
-        # pass in only source_ip
-        assert (
-            salt.transport.zeromq._get_master_uri(
-                master_ip=m_ip, master_port=m_port, source_ip=s_ip
-            )
-            == f"tcp://{s_ip}:0;{m_ip}:{m_port}"
-        )
-
-        assert (
-            salt.transport.zeromq._get_master_uri(
-                master_ip=m_ip6, master_port=m_port, source_ip=s_ip6
-            )
-            == f"tcp://[{s_ip6}]:0;[{m_ip6}]:{m_port}"
-        )
-
-        # pass in only source_port
-        assert (
-            salt.transport.zeromq._get_master_uri(
-                master_ip=m_ip, master_port=m_port, source_port=s_port
-            )
-            == f"tcp://0.0.0.0:{s_port};{m_ip}:{m_port}"
-        )
-
-
-def test_clear_req_channel_master_uri_override(temp_salt_minion, temp_salt_master):
-    """
-    ensure master_uri kwarg is respected
-    """
-    opts = temp_salt_minion.config.copy()
-    # minion_config should be 127.0.0.1, we want a different uri that still connects
-    opts.update(
-        {
-            "id": "root",
-            "transport": "zeromq",
-            "auth_tries": 1,
-            "auth_timeout": 5,
-            "master_ip": "127.0.0.1",
-            "master_port": temp_salt_master.config["ret_port"],
-            "master_uri": "tcp://127.0.0.1:{}".format(
-                temp_salt_master.config["ret_port"]
-            ),
-        }
-    )
-    master_uri = "tcp://{master_ip}:{master_port}".format(
-        master_ip="localhost", master_port=opts["master_port"]
-    )
-    with salt.channel.client.ReqChannel.factory(opts, master_uri=master_uri) as channel:
-        assert "127.0.0.1" in channel.transport.message_client.addr
-
-
 def run_loop_in_thread(loop, evt):
     """
     Run the provided loop until an event is set
@@ -447,38 +353,6 @@ class MockSaltMinionMaster:
         """
         cls.mock._handle_payload_hook()
         raise tornado.gen.Return((payload, {"fun": "send_clear"}))
-
-
-@pytest.mark.parametrize("message", ["", [], ()])
-def test_badload(temp_salt_minion, temp_salt_master, message):
-    """
-    Test a variety of bad requests, make sure that we get some sort of error
-    """
-    with MockSaltMinionMaster(temp_salt_minion, temp_salt_master) as minion_master:
-        ret = minion_master.channel.send(message, timeout=5, tries=1)
-        assert ret == "payload and load must be a dict"
-
-
-def test_payload_handling_exception(temp_salt_minion, temp_salt_master):
-    """
-    test of getting exception on payload handling
-    """
-    with MockSaltMinionMaster(temp_salt_minion, temp_salt_master) as minion_master:
-        with patch.object(minion_master.mock, "_handle_payload_hook") as _mock:
-            _mock.side_effect = Exception()
-            ret = minion_master.channel.send({}, timeout=5, tries=1)
-            assert ret == "Some exception handling minion payload"
-
-
-def test_serverside_exception(temp_salt_minion, temp_salt_master):
-    """
-    test of getting server side exception on payload handling
-    """
-    with MockSaltMinionMaster(temp_salt_minion, temp_salt_master) as minion_master:
-        with patch.object(minion_master.mock, "_handle_payload_hook") as _mock:
-            _mock.side_effect = tornado.gen.Return(({}, {"fun": "madeup-fun"}))
-            ret = minion_master.channel.send({}, timeout=5, tries=1)
-            assert ret == "Server-side exception handling payload"
 
 
 def test_zeromq_async_pub_channel_publish_port(temp_salt_master):
@@ -600,934 +474,6 @@ def test_zeromq_async_pub_channel_filtering_decode_message(
     assert res.result()["enc"] == "aes"
 
 
-def test_req_server_chan_encrypt_v2(
-    pki_dir, encryption_algorithm, signing_algorithm, master_opts
-):
-    loop = tornado.ioloop.IOLoop.current()
-    master_opts.update(
-        {
-            "worker_threads": 1,
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "zmq_monitor": False,
-            "mworker_queue_niceness": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("master")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-        }
-    )
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    dictkey = "pillar"
-    nonce = "abcdefg"
-    pillar_data = {"pillar1": "meh"}
-    ret = server._encrypt_private(
-        pillar_data,
-        dictkey,
-        "minion",
-        nonce,
-        encryption_algorithm=encryption_algorithm,
-        signing_algorithm=signing_algorithm,
-    )
-    assert "key" in ret
-    assert dictkey in ret
-
-    key = salt.crypt.PrivateKey(str(pki_dir.joinpath("minion", "minion.pem")))
-    aes = key.decrypt(ret["key"], encryption_algorithm)
-    pcrypt = salt.crypt.Crypticle(master_opts, aes)
-    signed_msg = pcrypt.loads(ret[dictkey])
-
-    assert "sig" in signed_msg
-    assert "data" in signed_msg
-    data = salt.payload.loads(signed_msg["data"])
-    assert "key" in data
-    assert data["key"] == ret["key"]
-    assert "key" in data
-    assert data["nonce"] == nonce
-    assert "pillar" in data
-    assert data["pillar"] == pillar_data
-
-
-def test_req_server_chan_encrypt_v1(pki_dir, encryption_algorithm, master_opts):
-    loop = tornado.ioloop.IOLoop.current()
-    master_opts.update(
-        {
-            "worker_threads": 1,
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "zmq_monitor": False,
-            "mworker_queue_niceness": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("master")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-        }
-    )
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    dictkey = "pillar"
-    nonce = "abcdefg"
-    pillar_data = {"pillar1": "meh"}
-    ret = server._encrypt_private(
-        pillar_data,
-        dictkey,
-        "minion",
-        sign_messages=False,
-        encryption_algorithm=encryption_algorithm,
-    )
-
-    assert "key" in ret
-    assert dictkey in ret
-
-    key = salt.crypt.PrivateKey(str(pki_dir.joinpath("minion", "minion.pem")))
-    aes = key.decrypt(ret["key"], encryption_algorithm)
-    pcrypt = salt.crypt.Crypticle(master_opts, aes)
-    data = pcrypt.loads(ret[dictkey])
-    assert data == pillar_data
-
-
-def test_req_chan_decode_data_dict_entry_v1(
-    pki_dir, encryption_algorithm, minion_opts, master_opts
-):
-    mockloop = MagicMock()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    master_opts = dict(master_opts, pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    client = salt.channel.client.ReqChannel.factory(minion_opts, io_loop=mockloop)
-    dictkey = "pillar"
-    target = "minion"
-    pillar_data = {"pillar1": "meh"}
-    ret = server._encrypt_private(
-        pillar_data,
-        dictkey,
-        target,
-        sign_messages=False,
-        encryption_algorithm=encryption_algorithm,
-    )
-    key = client.auth.get_keys()
-    aes = key.decrypt(ret["key"], encryption_algorithm)
-    pcrypt = salt.crypt.Crypticle(client.opts, aes)
-    ret_pillar_data = pcrypt.loads(ret[dictkey])
-    assert ret_pillar_data == pillar_data
-
-
-async def test_req_chan_decode_data_dict_entry_v2(minion_opts, master_opts, pki_dir):
-    mockloop = MagicMock()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=mockloop)
-
-    dictkey = "pillar"
-    target = "minion"
-    pillar_data = {"pillar1": "meh"}
-
-    # Mock auth and message client.
-    auth = client.auth
-    auth._crypticle = salt.crypt.Crypticle(minion_opts, AES_KEY)
-    client.auth = MagicMock()
-    client.auth.mpub = auth.mpub
-    client.auth.authenticated = True
-    client.auth.get_keys = auth.get_keys
-    client.auth.crypticle.dumps = auth.crypticle.dumps
-    client.auth.crypticle.loads = auth.crypticle.loads
-    client.transport = MagicMock()
-
-    print(minion_opts["encryption_algorithm"])
-
-    @tornado.gen.coroutine
-    def mocksend(msg, timeout=60, tries=3):
-        client.transport.msg = msg
-        load = client.auth.crypticle.loads(msg["load"])
-        ret = server._encrypt_private(
-            pillar_data,
-            dictkey,
-            target,
-            nonce=load["nonce"],
-            sign_messages=True,
-            encryption_algorithm=minion_opts["encryption_algorithm"],
-            signing_algorithm=minion_opts["signing_algorithm"],
-        )
-        raise tornado.gen.Return(ret)
-
-    client.transport.send = mocksend
-
-    # Note the 'ver' value in 'load' does not represent the the 'version' sent
-    # in the top level of the transport's message.
-    load = {
-        "id": target,
-        "grains": {},
-        "saltenv": "base",
-        "pillarenv": "base",
-        "pillar_override": True,
-        "extra_minion_data": {},
-        "ver": "2",
-        "cmd": "_pillar",
-    }
-    ret = await client.crypted_transfer_decode_dictentry(
-        load,
-        dictkey="pillar",
-    )
-    assert "version" in client.transport.msg
-    assert client.transport.msg["version"] == 2
-    assert ret == {"pillar1": "meh"}
-
-
-async def test_req_chan_decode_data_dict_entry_v2_bad_nonce(
-    pki_dir, minion_opts, master_opts
-):
-    mockloop = MagicMock()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=mockloop)
-
-    dictkey = "pillar"
-    badnonce = "abcdefg"
-    target = "minion"
-    pillar_data = {"pillar1": "meh"}
-
-    # Mock auth and message client.
-    auth = client.auth
-    auth._crypticle = salt.crypt.Crypticle(minion_opts, AES_KEY)
-    client.auth = MagicMock()
-    client.auth.mpub = auth.mpub
-    client.auth.authenticated = True
-    client.auth.get_keys = auth.get_keys
-    client.auth.crypticle.dumps = auth.crypticle.dumps
-    client.auth.crypticle.loads = auth.crypticle.loads
-    client.transport = MagicMock()
-    ret = server._encrypt_private(
-        pillar_data,
-        dictkey,
-        target,
-        nonce=badnonce,
-        sign_messages=True,
-        encryption_algorithm=minion_opts["encryption_algorithm"],
-        signing_algorithm=minion_opts["signing_algorithm"],
-    )
-
-    @tornado.gen.coroutine
-    def mocksend(msg, timeout=60, tries=3):
-        client.transport.msg = msg
-        raise tornado.gen.Return(ret)
-
-    client.transport.send = mocksend
-
-    # Note the 'ver' value in 'load' does not represent the the 'version' sent
-    # in the top level of the transport's message.
-    load = {
-        "id": target,
-        "grains": {},
-        "saltenv": "base",
-        "pillarenv": "base",
-        "pillar_override": True,
-        "extra_minion_data": {},
-        "ver": "2",
-        "cmd": "_pillar",
-    }
-
-    with pytest.raises(salt.crypt.AuthenticationError) as excinfo:
-        ret = await client.crypted_transfer_decode_dictentry(
-            load,
-            dictkey="pillar",
-        )
-    assert "Pillar nonce verification failed." == excinfo.value.message
-
-
-async def test_req_chan_decode_data_dict_entry_v2_bad_signature(
-    pki_dir, minion_opts, master_opts
-):
-    mockloop = MagicMock()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=mockloop)
-
-    dictkey = "pillar"
-    badnonce = "abcdefg"
-    target = "minion"
-    pillar_data = {"pillar1": "meh"}
-
-    # Mock auth and message client.
-    auth = client.auth
-    auth._crypticle = salt.crypt.Crypticle(minion_opts, AES_KEY)
-    client.auth = MagicMock()
-    client.auth.mpub = auth.mpub
-    client.auth.authenticated = True
-    client.auth.get_keys = auth.get_keys
-    client.auth.crypticle.dumps = auth.crypticle.dumps
-    client.auth.crypticle.loads = auth.crypticle.loads
-    client.transport = MagicMock()
-
-    @tornado.gen.coroutine
-    def mocksend(msg, timeout=60, tries=3):
-        client.transport.msg = msg
-        load = client.auth.crypticle.loads(msg["load"])
-        ret = server._encrypt_private(
-            pillar_data,
-            dictkey,
-            target,
-            nonce=load["nonce"],
-            sign_messages=True,
-            encryption_algorithm=minion_opts["encryption_algorithm"],
-            signing_algorithm=minion_opts["signing_algorithm"],
-        )
-
-        key = client.auth.get_keys()
-        aes = key.decrypt(ret["key"], minion_opts["encryption_algorithm"])
-        pcrypt = salt.crypt.Crypticle(client.opts, aes)
-        signed_msg = pcrypt.loads(ret[dictkey])
-        # Changing the pillar data will cause the signature verification to
-        # fail.
-        data = salt.payload.loads(signed_msg["data"])
-        data["pillar"] = {"pillar1": "bar"}
-        signed_msg["data"] = salt.payload.dumps(data)
-        ret[dictkey] = pcrypt.dumps(signed_msg)
-        raise tornado.gen.Return(ret)
-
-    client.transport.send = mocksend
-
-    # Note the 'ver' value in 'load' does not represent the the 'version' sent
-    # in the top level of the transport's message.
-    load = {
-        "id": target,
-        "grains": {},
-        "saltenv": "base",
-        "pillarenv": "base",
-        "pillar_override": True,
-        "extra_minion_data": {},
-        "ver": "2",
-        "cmd": "_pillar",
-    }
-
-    with pytest.raises(salt.crypt.AuthenticationError) as excinfo:
-        ret = await client.crypted_transfer_decode_dictentry(
-            load,
-            dictkey="pillar",
-        )
-    assert "Pillar payload signature failed to validate." == excinfo.value.message
-
-
-async def test_req_chan_decode_data_dict_entry_v2_bad_key(
-    pki_dir, minion_opts, master_opts
-):
-    mockloop = MagicMock()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=mockloop)
-
-    dictkey = "pillar"
-    badnonce = "abcdefg"
-    target = "minion"
-    pillar_data = {"pillar1": "meh"}
-
-    # Mock auth and message client.
-    auth = client.auth
-    auth._crypticle = salt.crypt.Crypticle(master_opts, AES_KEY)
-    client.auth = MagicMock()
-    client.auth.mpub = auth.mpub
-    client.auth.authenticated = True
-    client.auth.get_keys = auth.get_keys
-    client.auth.crypticle.dumps = auth.crypticle.dumps
-    client.auth.crypticle.loads = auth.crypticle.loads
-    client.transport = MagicMock()
-
-    @tornado.gen.coroutine
-    def mocksend(msg, timeout=60, tries=3):
-        client.transport.msg = msg
-        load = client.auth.crypticle.loads(msg["load"])
-        ret = server._encrypt_private(
-            pillar_data,
-            dictkey,
-            target,
-            nonce=load["nonce"],
-            sign_messages=True,
-            encryption_algorithm=minion_opts["encryption_algorithm"],
-            signing_algorithm=minion_opts["signing_algorithm"],
-        )
-
-        mkey = client.auth.get_keys()
-        aes = mkey.decrypt(ret["key"], minion_opts["encryption_algorithm"])
-        pcrypt = salt.crypt.Crypticle(client.opts, aes)
-        signed_msg = pcrypt.loads(ret[dictkey])
-
-        # Now encrypt with a different key
-        key = salt.crypt.Crypticle.generate_key_string()
-        pcrypt = salt.crypt.Crypticle(master_opts, key)
-        pubfn = os.path.join(master_opts["pki_dir"], "minions", "minion")
-        pub = salt.crypt.PublicKey(pubfn)
-        ret[dictkey] = pcrypt.dumps(signed_msg)
-        key = salt.utils.stringutils.to_bytes(key)
-        ret["key"] = pub.encrypt(key, minion_opts["encryption_algorithm"])
-        raise tornado.gen.Return(ret)
-
-    client.transport.send = mocksend
-
-    # Note the 'ver' value in 'load' does not represent the the 'version' sent
-    # in the top level of the transport's message.
-    load = {
-        "id": target,
-        "grains": {},
-        "saltenv": "base",
-        "pillarenv": "base",
-        "pillar_override": True,
-        "extra_minion_data": {},
-        "ver": "2",
-        "cmd": "_pillar",
-    }
-    try:
-        with pytest.raises(salt.crypt.AuthenticationError) as excinfo:
-            await client.crypted_transfer_decode_dictentry(
-                load,
-                dictkey="pillar",
-            )
-        assert "Key verification failed." == excinfo.value.message
-    finally:
-        client.close()
-        server.close()
-
-
-async def test_req_serv_auth_v1(pki_dir, minion_opts, master_opts):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "master_sign_pubkey": False,
-            "publish_port": 4505,
-            "auth_mode": 1,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-
-    pub = salt.crypt.get_rsa_pub_key(str(pki_dir.joinpath("minion", "minion.pub")))
-    token = salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string())
-    nonce = uuid.uuid4().hex
-
-    # We need to read the public key with fopen otherwise the newlines might
-    # not match on windows.
-    with salt.utils.files.fopen(
-        str(pki_dir.joinpath("minion", "minion.pub")), "r"
-    ) as fp:
-        pub_key = salt.crypt.clean_key(fp.read())
-
-    load = {
-        "cmd": "_auth",
-        "id": "minion",
-        "token": token,
-        "pub": pub_key,
-        "enc_algo": minion_opts["encryption_algorithm"],
-        "sig_algo": minion_opts["signing_algorithm"],
-    }
-    ret = server._auth(load, sign_messages=False)
-    assert "load" not in ret
-
-
-async def test_req_serv_auth_v2(pki_dir, minion_opts, master_opts):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "master_sign_pubkey": False,
-            "publish_port": 4505,
-            "auth_mode": 1,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-
-    pub = salt.crypt.get_rsa_pub_key(str(pki_dir.joinpath("minion", "minion.pub")))
-    token = salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string())
-    nonce = uuid.uuid4().hex
-
-    # We need to read the public key with fopen otherwise the newlines might
-    # not match on windows.
-    with salt.utils.files.fopen(
-        str(pki_dir.joinpath("minion", "minion.pub")), "r"
-    ) as fp:
-        pub_key = fp.read()
-
-    load = {
-        "cmd": "_auth",
-        "id": "minion",
-        "nonce": nonce,
-        "token": token,
-        "pub": pub_key,
-        "enc_algo": minion_opts["encryption_algorithm"],
-        "sig_algo": minion_opts["signing_algorithm"],
-    }
-    ret = server._auth(load, sign_messages=True)
-    assert "sig" in ret
-    assert "load" in ret
-
-
-async def test_req_chan_auth_v2(pki_dir, io_loop, minion_opts, master_opts):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "publish_port": 4505,
-            "auth_mode": 1,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    master_opts["master_sign_pubkey"] = False
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    minion_opts["verify_master_pubkey_sign"] = False
-    minion_opts["always_verify_signature"] = False
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=io_loop)
-    signin_payload = client.auth.minion_sign_in_payload()
-    pload = client._package_load(signin_payload)
-    assert "version" in pload
-    assert pload["version"] == 2
-
-    ret = server._auth(pload["load"], sign_messages=True)
-    assert "sig" in ret
-    ret = client.auth.handle_signin_response(signin_payload, ret)
-    assert "aes" in ret
-    assert "master_uri" in ret
-    assert "publish_port" in ret
-
-
-async def test_req_chan_auth_v2_with_master_signing(
-    pki_dir, io_loop, minion_opts, master_opts
-):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "publish_port": 4505,
-            "auth_mode": 1,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts = dict(master_opts, pki_dir=str(pki_dir.joinpath("master")))
-    master_opts["master_sign_pubkey"] = True
-    master_opts["master_use_pubkey_signature"] = False
-    master_opts["signing_key_pass"] = ""
-    master_opts["master_sign_key_name"] = "master_sign"
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    minion_opts["verify_master_pubkey_sign"] = True
-    minion_opts["always_verify_signature"] = True
-    minion_opts["master_sign_key_name"] = "master_sign"
-    minion_opts["master"] = "master"
-
-    assert (
-        pki_dir.joinpath("minion", "minion_master.pub").read_text()
-        == pki_dir.joinpath("master", "master.pub").read_text()
-    )
-
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=io_loop)
-    signin_payload = client.auth.minion_sign_in_payload()
-    pload = client._package_load(signin_payload)
-    assert "version" in pload
-    assert pload["version"] == 2
-
-    server_reply = server._auth(pload["load"], sign_messages=True)
-    # With version 2 we always get a clear signed response
-    assert "enc" in server_reply
-    assert server_reply["enc"] == "clear"
-    assert "sig" in server_reply
-    assert "load" in server_reply
-    ret = client.auth.handle_signin_response(signin_payload, server_reply)
-    assert "aes" in ret
-    assert "master_uri" in ret
-    assert "publish_port" in ret
-
-    # Now create a new master key pair and try auth with it.
-    mapriv = pki_dir.joinpath("master", "master.pem")
-    mapriv.unlink()
-    mapriv.write_text(MASTER2_PRIV_KEY.strip())
-    mapub = pki_dir.joinpath("master", "master.pub")
-    mapub.unlink()
-    mapub.write_text(MASTER2_PUB_KEY.strip())
-
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-
-    signin_payload = client.auth.minion_sign_in_payload()
-    pload = client._package_load(signin_payload)
-    server_reply = server._auth(pload["load"], sign_messages=True)
-    ret = client.auth.handle_signin_response(signin_payload, server_reply)
-
-    assert "aes" in ret
-    assert "master_uri" in ret
-    assert "publish_port" in ret
-
-    assert (
-        pki_dir.joinpath("minion", "minion_master.pub").read_text()
-        == pki_dir.joinpath("master", "master.pub").read_text()
-    )
-
-
-async def test_req_chan_auth_v2_new_minion_with_master_pub(
-    pki_dir, io_loop, minion_opts, master_opts
-):
-
-    pki_dir.joinpath("master", "minions", "minion").unlink()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "publish_port": 4505,
-            "auth_mode": 1,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    master_opts["master_sign_pubkey"] = False
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    minion_opts["verify_master_pubkey_sign"] = False
-    minion_opts["always_verify_signature"] = False
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=io_loop)
-    signin_payload = client.auth.minion_sign_in_payload()
-    pload = client._package_load(signin_payload)
-    assert "version" in pload
-    assert pload["version"] == 2
-
-    ret = server._auth(pload["load"], sign_messages=True)
-    assert "sig" in ret
-    ret = client.auth.handle_signin_response(signin_payload, ret)
-    assert ret == "retry"
-
-
-async def test_req_chan_auth_v2_new_minion_with_master_pub_bad_sig(
-    pki_dir, io_loop, minion_opts, master_opts
-):
-
-    pki_dir.joinpath("master", "minions", "minion").unlink()
-
-    # Give the master a different key than the minion has.
-    mapriv = pki_dir.joinpath("master", "master.pem")
-    mapriv.unlink()
-    mapriv.write_text(MASTER2_PRIV_KEY.strip())
-    mapub = pki_dir.joinpath("master", "master.pub")
-    mapub.unlink()
-    mapub.write_text(MASTER2_PUB_KEY.strip())
-
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "publish_port": 4505,
-            "auth_mode": 1,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(
-        pki_dir=str(pki_dir.joinpath("master")), master_sign_pubkey=False
-    )
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    minion_opts["verify_master_pubkey_sign"] = False
-    minion_opts["always_verify_signature"] = False
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=io_loop)
-    signin_payload = client.auth.minion_sign_in_payload()
-    pload = client._package_load(signin_payload)
-    assert "version" in pload
-    assert pload["version"] == 2
-
-    ret = server._auth(pload["load"], sign_messages=True)
-    assert "sig" in ret
-    with pytest.raises(salt.crypt.SaltClientError, match="Invalid signature"):
-        ret = client.auth.handle_signin_response(signin_payload, ret)
-
-
-async def test_req_chan_auth_v2_new_minion_without_master_pub(
-    minion_opts,
-    master_opts,
-    pki_dir,
-    io_loop,
-):
-
-    pki_dir.joinpath("master", "minions", "minion").unlink()
-    pki_dir.joinpath("minion", "minion_master.pub").unlink()
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "publish_port": 4505,
-            "auth_mode": 1,
-            "acceptance_wait_time": 3,
-            "acceptance_wait_time_max": 3,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    master_opts["master_sign_pubkey"] = False
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    minion_opts["verify_master_pubkey_sign"] = False
-    minion_opts["always_verify_signature"] = False
-    client = salt.channel.client.AsyncReqChannel.factory(minion_opts, io_loop=io_loop)
-    signin_payload = client.auth.minion_sign_in_payload()
-    pload = client._package_load(signin_payload)
-    try:
-        assert "version" in pload
-        assert pload["version"] == 2
-
-        ret = server._auth(pload["load"], sign_messages=True)
-        assert "sig" in ret
-        ret = client.auth.handle_signin_response(signin_payload, ret)
-        assert ret == "retry"
-    finally:
-        client.close()
-        server.close()
-
-
 async def test_req_server_garbage_request(io_loop):
     """
     Validate invalid msgpack messages will not raise exceptions in the
@@ -1639,303 +585,3 @@ async def test_unclosed_publish_client(minion_opts, io_loop):
             client.__del__()  # pylint: disable=unnecessary-dunder-call
     finally:
         client.close()
-
-
-@pytest.mark.skipif(not FIPS_TESTRUN, reason="Only run on fips enabled platforms")
-def test_req_server_auth_unsupported_sig_algo(
-    pki_dir, minion_opts, master_opts, caplog
-):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "master_sign_pubkey": False,
-            "publish_port": 4505,
-            "auth_mode": 1,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    pub = salt.crypt.PublicKey(str(pki_dir.joinpath("master", "master.pub")))
-    token = pub.encrypt(
-        salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        algorithm=minion_opts["encryption_algorithm"],
-    )
-    nonce = uuid.uuid4().hex
-
-    # We need to read the public key with fopen otherwise the newlines might
-    # not match on windows.
-    with salt.utils.files.fopen(
-        str(pki_dir.joinpath("minion", "minion.pub")), "r"
-    ) as fp:
-        pub_key = salt.crypt.clean_key(fp.read())
-
-    load = {
-        "version": 2,
-        "cmd": "_auth",
-        "id": "minion",
-        "token": token,
-        "pub": pub_key,
-        "nonce": "asdfse",
-        "enc_algo": minion_opts["encryption_algorithm"],
-        "sig_algo": salt.crypt.PKCS1v15_SHA1,
-    }
-    with caplog.at_level(logging.INFO):
-        ret = server._auth(load, sign_messages=True)
-        assert (
-            "Minion tried to authenticate with unsupported signing algorithm: PKCS1v15-SHA1"
-            in caplog.text
-        )
-        assert "load" in ret
-        assert "ret" in ret["load"]
-        assert ret["load"]["ret"] == "bad sig algo"
-
-
-def test_req_server_auth_garbage_sig_algo(pki_dir, minion_opts, master_opts, caplog):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "master_sign_pubkey": False,
-            "publish_port": 4505,
-            "auth_mode": 1,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    pub = salt.crypt.PublicKey(str(pki_dir.joinpath("master", "master.pub")))
-    token = pub.encrypt(
-        salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        algorithm=minion_opts["encryption_algorithm"],
-    )
-    nonce = uuid.uuid4().hex
-
-    # We need to read the public key with fopen otherwise the newlines might
-    # not match on windows.
-    with salt.utils.files.fopen(
-        str(pki_dir.joinpath("minion", "minion.pub")), "r"
-    ) as fp:
-        pub_key = salt.crypt.clean_key(fp.read())
-
-    load = {
-        "version": 2,
-        "cmd": "_auth",
-        "id": "minion",
-        "token": token,
-        "pub": pub_key,
-        "nonce": "asdfse",
-        "enc_algo": minion_opts["encryption_algorithm"],
-        "sig_algo": "IAMNOTANALGO",
-    }
-    with caplog.at_level(logging.INFO):
-        ret = server._auth(load, sign_messages=True)
-        assert (
-            "Minion tried to authenticate with unsupported signing algorithm: IAMNOTANALGO"
-            in caplog.text
-        )
-        assert "load" in ret
-        assert "ret" in ret["load"]
-        assert ret["load"]["ret"] == "bad sig algo"
-
-
-@pytest.mark.skipif(not FIPS_TESTRUN, reason="Only run on fips enabled platforms")
-def test_req_server_auth_unsupported_enc_algo(
-    pki_dir, minion_opts, master_opts, caplog
-):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "master_sign_pubkey": False,
-            "publish_port": 4505,
-            "auth_mode": 1,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    import tests.pytests.unit.crypt
-
-    pub = tests.pytests.unit.crypt.LegacyPublicKey(
-        str(pki_dir.joinpath("master", "master.pub"))
-    )
-    token = pub.encrypt(
-        salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-    )
-    nonce = uuid.uuid4().hex
-
-    # We need to read the public key with fopen otherwise the newlines might
-    # not match on windows.
-    with salt.utils.files.fopen(
-        str(pki_dir.joinpath("minion", "minion.pub")), "r"
-    ) as fp:
-        pub_key = salt.crypt.clean_key(fp.read())
-
-    load = {
-        "version": 2,
-        "cmd": "_auth",
-        "id": "minion",
-        "token": token,
-        "pub": pub_key,
-        "nonce": "asdfse",
-        "enc_algo": "OAEP-SHA1",
-        "sig_algo": minion_opts["signing_algorithm"],
-    }
-    with caplog.at_level(logging.INFO):
-        ret = server._auth(load, sign_messages=True)
-        assert (
-            "Minion minion tried to authenticate with unsupported encryption algorithm: OAEP-SHA1"
-            in caplog.text
-        )
-        assert "load" in ret
-        assert "ret" in ret["load"]
-        assert ret["load"]["ret"] == "bad enc algo"
-
-
-def test_req_server_auth_garbage_enc_algo(pki_dir, minion_opts, master_opts, caplog):
-    minion_opts.update(
-        {
-            "master_uri": "tcp://127.0.0.1:4506",
-            "interface": "127.0.0.1",
-            "ret_port": 4506,
-            "ipv6": False,
-            "sock_dir": ".",
-            "pki_dir": str(pki_dir.joinpath("minion")),
-            "id": "minion",
-            "__role": "minion",
-            "keysize": 4096,
-            "max_minions": 0,
-            "auto_accept": False,
-            "open_mode": False,
-            "key_pass": None,
-            "master_sign_pubkey": False,
-            "publish_port": 4505,
-            "auth_mode": 1,
-        }
-    )
-    SMaster.secrets["aes"] = {
-        "secret": multiprocessing.Array(
-            ctypes.c_char,
-            salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-        ),
-        "reload": salt.crypt.Crypticle.generate_key_string,
-    }
-    master_opts.update(pki_dir=str(pki_dir.joinpath("master")))
-    server = salt.channel.server.ReqServerChannel.factory(master_opts)
-
-    server.auto_key = salt.daemons.masterapi.AutoKey(server.opts)
-    server.cache_cli = False
-    server.event = salt.utils.event.get_master_event(
-        master_opts, master_opts["sock_dir"], listen=False
-    )
-    server.master_key = salt.crypt.MasterKeys(server.opts)
-    import tests.pytests.unit.crypt
-
-    pub = tests.pytests.unit.crypt.LegacyPublicKey(
-        str(pki_dir.joinpath("master", "master.pub"))
-    )
-    token = pub.encrypt(
-        salt.utils.stringutils.to_bytes(salt.crypt.Crypticle.generate_key_string()),
-    )
-    nonce = uuid.uuid4().hex
-
-    # We need to read the public key with fopen otherwise the newlines might
-    # not match on windows.
-    with salt.utils.files.fopen(
-        str(pki_dir.joinpath("minion", "minion.pub")), "r"
-    ) as fp:
-        pub_key = salt.crypt.clean_key(fp.read())
-
-    load = {
-        "version": 2,
-        "cmd": "_auth",
-        "id": "minion",
-        "token": token,
-        "pub": pub_key,
-        "nonce": "asdfse",
-        "enc_algo": "IAMNOTAENCALGO",
-        "sig_algo": minion_opts["signing_algorithm"],
-    }
-    with caplog.at_level(logging.INFO):
-        ret = server._auth(load, sign_messages=True)
-        assert (
-            "Minion minion tried to authenticate with unsupported encryption algorithm: IAMNOTAENCALGO"
-            in caplog.text
-        )
-        assert "load" in ret
-        assert "ret" in ret["load"]
-        assert ret["load"]["ret"] == "bad enc algo"
